kvm: follow up to #1525 issue
ssh keys setup moved to stage1, from stage0. data directory
(as requested by @krnowak) separated from stage0 rkt datadir.
sshd port removed from configuration (inetd mode use only).
