+++ stage1/usr_from_kvm/files/sshd_config
@@ -1,4 +1,3 @@
-Port 22
 Protocol 2
 HostKey /etc/ssh/ssh_host_rsa_key
 HostKey /etc/ssh/ssh_host_dsa_key
